Finalize eager functions
