+++ py-polars/src/functions/eager.rs
@@ -1,12 +1,68 @@
-use polars::time;
+use polars::{functions, time};
 use polars_core::datatypes::{TimeUnit, TimeZone};
-use polars_core::prelude::IntoSeries;
+use polars_core::prelude::{DataFrame, IntoSeries};
 use pyo3::prelude::*;
 
-use crate::conversion::Wrap;
+use crate::conversion::{get_df, get_series, Wrap};
 use crate::error::PyPolarsErr;
 use crate::prelude::{ClosedWindow, Duration};
-use crate::PySeries;
+use crate::{PyDataFrame, PySeries};
+
+#[pyfunction]
+pub fn concat_df(dfs: &PyAny, py: Python) -> PyResult<PyDataFrame> {
+    use polars_core::error::PolarsResult;
+    use polars_core::utils::rayon::prelude::*;
+
+    let mut iter = dfs.iter()?;
+    let first = iter.next().unwrap()?;
+
+    let first_rdf = get_df(first)?;
+    let identity_df = first_rdf.clear();
+
+    let mut rdfs: Vec<PolarsResult<DataFrame>> = vec![Ok(first_rdf)];
+
+    for item in iter {
+        let rdf = get_df(item?)?;
+        rdfs.push(Ok(rdf));
+    }
+
+    let identity = || Ok(identity_df.clone());
+
+    let df = py
+        .allow_threads(|| {
+            polars_core::POOL.install(|| {
+                rdfs.into_par_iter()
+                    .fold(identity, |acc: PolarsResult<DataFrame>, df| {
+                        let mut acc = acc?;
+                        acc.vstack_mut(&df?)?;
+                        Ok(acc)
+                    })
+                    .reduce(identity, |acc, df| {
+                        let mut acc = acc?;
+                        acc.vstack_mut(&df?)?;
+                        Ok(acc)
+                    })
+            })
+        })
+        .map_err(PyPolarsErr::from)?;
+
+    Ok(df.into())
+}
+
+#[pyfunction]
+pub fn concat_series(series: &PyAny) -> PyResult<PySeries> {
+    let mut iter = series.iter()?;
+    let first = iter.next().unwrap()?;
+
+    let mut s = get_series(first)?;
+
+    for res in iter {
+        let item = res?;
+        let item = get_series(item)?;
+        s.append(&item).map_err(PyPolarsErr::from)?;
+    }
+    Ok(s.into())
+}
 
 #[pyfunction]
 pub fn date_range(
@@ -29,4 +85,34 @@ pub fn date_range(
     )
     .map_err(PyPolarsErr::from)?;
     Ok(date_range.into_series().into())
+}
+
+#[pyfunction]
+pub fn diag_concat_df(dfs: &PyAny) -> PyResult<PyDataFrame> {
+    let iter = dfs.iter()?;
+
+    let dfs = iter
+        .map(|item| {
+            let item = item?;
+            get_df(item)
+        })
+        .collect::<PyResult<Vec<_>>>()?;
+
+    let df = functions::diag_concat_df(&dfs).map_err(PyPolarsErr::from)?;
+    Ok(df.into())
+}
+
+#[pyfunction]
+pub fn hor_concat_df(dfs: &PyAny) -> PyResult<PyDataFrame> {
+    let iter = dfs.iter()?;
+
+    let dfs = iter
+        .map(|item| {
+            let item = item?;
+            get_df(item)
+        })
+        .collect::<PyResult<Vec<_>>>()?;
+
+    let df = functions::hor_concat_df(&dfs).map_err(PyPolarsErr::from)?;
+    Ok(df.into())
 }
